+++ src/index.ts
@@ -6,6 +6,9 @@ export * from "./permissions";
 export * from "./live_chat";
 export * from "./health";
 export * from "./ticket";
+export * from "./companies";
+export * from "./products";
+export * from "./chatbot-ura";
 
 export { config } from "./config";
 export type { LibraryConfig } from "./config";