+++ CHANGELOG.md
@@ -22,6 +22,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+# 0.7.0 - 2026-04-30
+
+### Added
+
+- Chatbot integration ([#93](https://github.com/orgs/Titus-System/projects/11/views/1?filterQuery=sprints%3A%22Sprints+2%22&sliceBy%5Bvalue%5D=WesleyGoncalves&pane=issue&itemId=176735264&issue=Titus-System%7CSyncDesk%7C93))
+- Update Tickets endpoints for new API version ([#94](https://github.com/orgs/Titus-System/projects/11/views/1?filterQuery=sprints%3A%22Sprints+2%22&sliceBy%5Bvalue%5D=WesleyGoncalves&pane=issue&itemId=176735629&issue=Titus-System%7CSyncDesk%7C94))
+- Atualização de Contratos e Tipagem([#90](https://github.com/Titus-System/SyncDesk/issues/90))
+- Integrar endpoints de Empresas	([#92](https://github.com/Titus-System/SyncDesk/issues/92))
+- Integrar endpoints de Produtos	([#91](https://github.com/Titus-System/SyncDesk/issues/91))
+- Integrar endpoints do Chatbot	([#93](https://github.com/Titus-System/SyncDesk/issues/93))
+- Integrar endpoints de Tickets	([#94](https://github.com/Titus-System/SyncDesk/issues/94))
+
+# 0.6.0 - 2026-04-20
+
+### Added
+
+- Support the Companies API endpoints for managing company, its products and users.
+- Support the Products API endpoints for managing company, its products and users.
+
 # 0.5.0 - 2026-04-04
 
 ### Added